backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8; python_version >= '3'
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers==3.10.0
django-filter==21.1
django-rest-passwordreset==1.2.1
django==3.2.9
djangorestframework-simplejwt==5.0.0
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3; python_version >= '3'
importlib-metadata==4.8.2; python_version < '3.10'
inflection==0.5.1
itypes==1.2.0
jinja2==3.0.3
markdown==3.3.6
markupsafe==2.0.1
mysqlclient==2.1.0
packaging==21.3
pyjwt==2.1.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
ruamel.yaml.clib==0.2.6; platform_python_implementation == 'CPython' and python_version < '3.10'
ruamel.yaml==0.17.17
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
